+++ app/helpers/application_helper.rb
@@ -31,7 +31,7 @@ module ApplicationHelper
   end
 
   def enable_google_tag_manager?
-    return false unless ENV["SENTRY_ENV"] == "production"
+    return false unless ENV["USER_ENV"] == "production"
     return false unless ENV["GOOGLE_TAG_MANAGER_ID"].present?
     return false unless cookies[:ACCEPT_OPTIONAL_COOKIES] == "true"
     true