+++ app/helpers/environment_banner_helper.rb
@@ -6,16 +6,16 @@ module EnvironmentBannerHelper
   }.freeze.with_indifferent_access
 
   def environment_banner
-    return if ENVIRONMENT_COLOURS.keys.none?(sentry_env)
+    return if ENVIRONMENT_COLOURS.keys.none?(user_env)
 
     govuk_tag(
-      text: "#{sentry_env.tr("_", " ").upcase} ENVIRONMENT",
-      colour: ENVIRONMENT_COLOURS[sentry_env],
+      text: "#{user_env.tr("_", " ").upcase} ENVIRONMENT",
+      colour: ENVIRONMENT_COLOURS[user_env],
       classes: "environment-banner"
     )
   end
 
-  private def sentry_env
+  private def user_env
     ENV.fetch("USER_ENV", "local_development")
   end
 end